+++ tests/test_connector.py
@@ -52,4 +52,4 @@ class TestDBConnector:
 
 
 if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
+    pytest.main([__file__, "-svv"])