+++ tests/test_utils.py
@@ -306,4 +306,4 @@ class TestLogDiagram:
 
 
 if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
+    pytest.main([__file__, "-svv"])